move endpoints to routers
